Amélioration des graphs d'affichage

+++ group3_genetic/src/main/java/fr/isen/cir56/group3_genetic/Analyzer/AnalyzerInterface.java
@@ -22,7 +22,7 @@ public interface AnalyzerInterface {
 	 * Get the number of generations
 	 * @return 
 	 */
-	double getNumberGenerations() throws UnexistingBreederException;
+	int getNumberGenerations() throws UnexistingBreederException;
 	
 	/**
 	 * Get the percentage of involving for the generation compared to the previous one
